pyproject.toml
[project]
name = "brainstorm"
version = "0.1.0"
description = "AI-powered interesting project idea generator app"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "google-genai>=1.68.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
    "ty>=0.0.23",
]
